backend/backend/requirements.txt
Flask==2.3.2
Gunicorn==21.2.0
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.3
Flask-Cors==3.0.10

# Database
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0

# Blockchain (Web3.py)
web3==6.11.0

# KYC / Sumsub
requests==2.31.0
PyJWT==2.8.0

# MFA TOTP
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.0.0

# Email
sendgrid==6.11.0

# Redis (rate limiting)
redis==5.0.0

# Utilities
python-dateutil==2.8.2
pydantic==1.10.12

# Tests
pytest==7.4.2
pytest-cov==4.1.0
